Split backup code in 2 files
